celery_with_django/requirements.txt
faster-whisper
ffmpeg-python
pytube
fastapi[standard]
rq 
redis
celery[redis]
deep-translator
django
djangorestframework

celery -A django_celery_project.celery worker --pool=solo -l info
python manage.py runserver

http://localhost:8000/download/
{
    "url": "https://www.youtube.com/watch?v=t9CAFYn7YgY"
}

http://localhost:8000/generate/
{
    "yt_id":"t9CAFYn7YgY",
    "dest":"en"
}